Cleaned up slab reactor tests.
